requirements.txt
google-generativeai
pydantic
python-dotenv
Pillow
fastapi
uvicorn
gradio
httpx
pypdf
PyMuPDF
pdfplumber
openpyxl